+++ VERSIONHISTORY.md
@@ -1,6 +1,12 @@
 # Harley - Version history
 Updates by version
 
+### Version: 1.2.3
+
+Minor:
+- Model Training: Fixed labeling issue, where split polygons could not be removed easily by clicking.
+- Data Loader: UI Change for adding batches, batch creator and added option to apply pipeline parameters to all batches with one click.
+
 ### Version: 1.2.2
 
 Minor: